refactor: move Express app setup directly to index.ts, remove app factory
Architecture principle: 'no function wrapper rules' - keep setup inline in index.ts
where it's actually used, not in a separate factory.

Changes:
- Deleted src/app.ts (factory function)
- Moved all useExpressServer setup directly into src/index.ts
- Removed global app.use(express.json()) - not needed with @JsonController
- routing-controllers with @JsonController handles body parsing automatically
- Test app setup remains in test-helpers.ts for test isolation

Benefits:
- Single location for app configuration (index.ts)
- No unnecessary function wrappers
- @JsonController eliminates need for global body parsing
- Simpler, more direct code path
