+++ packages/backend/tests/test-helpers.ts
@@ -23,11 +23,41 @@ export interface TestUser {
 
 /**
  * Returns a fully configured Express app for testing.
- * Uses the same factory function as src/index.ts to ensure consistency.
+ * Sets up routing-controllers with same configuration as src/index.ts.
  */
 export async function createTestApp(): Promise<Express> {
-  const { createApp } = await import('../src/app');
-  return createApp();
+  const express = await import('express');
+  const { useExpressServer } = await import('routing-controllers');
+  const { authorizationChecker, currentUserChecker } = await import('../src/middleware/routingControllersAuth');
+  const { correlationIdMiddleware } = await import('../src/middleware/correlationId.middleware');
+  const { requestLoggingMiddleware } = await import('../src/middleware/requestLogging.middleware');
+  const { controllers } = await import('../src/controllers');
+  const { appConfig } = await import('../src/config/appConfig');
+
+  const testApp = express.default();
+
+  useExpressServer(testApp, {
+    controllers: controllers,
+    authorizationChecker: authorizationChecker,
+    currentUserChecker: currentUserChecker,
+    defaultErrorHandler: true,
+    validation: {
+      whitelist: true,
+      forbidNonWhitelisted: true,
+    },
+    classTransformer: true,
+    cors: {
+      origin: appConfig.APP_FRONTEND_URL,
+      credentials: true,
+      exposedHeaders: ['set-auth-token', 'x-total-count', 'x-current-page', 'x-total-pages'],
+    },
+    middlewares: [
+      correlationIdMiddleware,
+      requestLoggingMiddleware,
+    ],
+  });
+
+  return testApp as Express;
 }
 
 /**
